requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
openai==1.58.1
apscheduler==3.10.4
json-repair==0.30.2
